refactor: 抽出 useCountdownTick 共用倒數邏輯
將 usePomodoro 內的 rAF + Date.now delta 倒數抽成共用 hook，
供番茄鐘與後續計時器共用，消除重複。
